+++ cid/backends/mysql/base.py
@@ -5,6 +5,6 @@ from ...cursor import CidCursorWrapper
 
 class DatabaseWrapper(BaseMySQLWrapper):
 
-    def create_cursor(self):
-        base_cursor = super(DatabaseWrapper, self).create_cursor()
+    def create_cursor(self, name):
+        base_cursor = super(DatabaseWrapper, self).create_cursor(name)
         return CidCursorWrapper(base_cursor)