+++ HISTORY.rst
@@ -25,3 +25,4 @@ History
 * Drop support of Python 2.6.
 * Generate cid outside of the middleware when ``GENERATE_CID`` is
   enabled, so that it's available even if the middleware is not used
+* Fix support of Django 1.11 in database backends.